Use the upstream json-reformat.el, not a local copy
The official home of `json-reformat` is https://github.com/gongo/json-reformat,
and MELPA has packages of that library. It is therefore undesirable to
distribute a copy with `restclient`.

This commit declares (and documents) a package dependency so that
everything will work correctly when installing the restclient package,
and it removes the local copy.