Added ability narrow down to current request
Let's see if it's going to be of any use. Currently it's not quite up to
org-mode's "edit special" level of quality, i.e. it doesn't really
create a separate buffer with proper major mode applied and ability to
commit or cancel the change. Instead it just does usually confusing
narrow-to-region with "current" boundaries detected, thus requiring to
press `C-x n w` to exit.