add hook after response received, before prettify
This adds a 3rd hook, `response-received-hook`, that runs within the response
buffer, but before the response is prettified. The existing hook
`response-loaded-hook` runs after the response is prettified. Together, both
hooks provide finer control over the response output: functions in
`response-received` operate on the raw response, and functions in
`response-loaded` operate on the formatted response.